test: update black hole test cases
