+++ vm/contract/blackhole.go
@@ -70,23 +70,6 @@ func (b *BlackHole) ProcessSend(ctx *vmstore.VMContext, block *types.StateBlock)
 	}
 }
 
-//func (b *BlackHole) DoPending(block *types.StateBlock) (*types.PendingKey, *types.PendingInfo, error) {
-//	param := new(cabi.DestroyParam)
-//	err := cabi.BlackHoleABI.UnpackMethod(param, cabi.MethodNameDestroy, block.Data)
-//	if err != nil {
-//		return nil, nil, err
-//	}
-//
-//	return &types.PendingKey{
-//			Address: param.Owner,
-//			Hash:    block.GetHash(),
-//		}, &types.PendingInfo{
-//			Source: types.Address(block.Link),
-//			Amount: types.ZeroBalance,
-//			Type:   block.Token,
-//		}, nil
-//}
-
 func (b *BlackHole) verify(ctx *vmstore.VMContext, param *cabi.DestroyParam, block *types.StateBlock) error {
 	if verify, err := param.Verify(); err != nil {
 		return err
